Extract shared code in base class
